Moved the default PATH setting to containerizer.
Currently, it is set as part of the executor environment. This is not
ideal because isolators might choose to set environment variables for
the contianer. We should set the default PATH in the last step (right
before launch) if it does not exist.

Mesos containerizer launch helper already does default PATH setting
right before launching the container. This patch did the same in
Docker containerizer, and remove the default PATH setting when
generating the executor environment.

Review: https://reviews.apache.org/r/54102