Added the initial implementation for applying offer operations.
The resource provider manager provides an `applyOfferOperation` method
for offer operation affecting resource providers. The resources on
which the operation should be applied contains a resource provider ID.
This will be extracted and an event will be sent to the respective
resource provider.

(This is based on https://reviews.apache.org/r/61810)

Review: https://reviews.apache.org/r/63480